@@ -6,12 +6,12 @@
     "start": "ng serve",
     "build": "ng build",
     "test": "ng test",
+    "test:lib": "ng test --project ngb-filterable-dropdown --watch false",
+    "test:sample": "ng test --project ngb-filterable-dropdown-example --watch false",
     "pretest": "npm run lint",
     "lint": "ng lint --fix",
-    "e2e": "ng e2e",
-    "lib:build": "ng build ngb-filterable-dropdown",
-    "lib:build-prod": "ng build ngb-filterable-dropdown --configuration production && cp README.md LICENSE.md dist/@bugsplat/ngb-filterable-dropdown",
-    "lib:release": "npm run lib:build-prod && cd dist/@bugsplat/ngb-filterable-dropdown && npm publish"
+    "build:lib": "ng build ngb-filterable-dropdown --configuration production && copyfiles README.md LICENSE.md dist/@bugsplat/ngb-filterable-dropdown",
+    "release": "npm run build:lib && cd dist/@bugsplat/ngb-filterable-dropdown && npm publish"
   },
   "private": false,
   "license": "MIT",
@@ -23,37 +23,38 @@
     "url": "https://github.com/BugSplat-Git/ngb-filterable-dropdown/issues"
   },
   "dependencies": {
-    "@angular/animations": "~13.3.0",
-    "@angular/common": "~13.3.0",
-    "@angular/compiler": "~13.3.0",
-    "@angular/core": "~13.3.0",
-    "@angular/forms": "~13.3.0",
-    "@angular/platform-browser": "~13.3.0",
-    "@angular/platform-browser-dynamic": "~13.3.0",
-    "@angular/router": "~13.3.0",
-    "@ng-bootstrap/ng-bootstrap": "^11.0.0",
-    "bootstrap": "^4.6.1",
-    "popper.js": "^1.16.1",
-    "rxjs": "~6.6.3",
+    "@angular/animations": "^14.2.1",
+    "@angular/common": "^14.2.1",
+    "@angular/compiler": "^14.2.1",
+    "@angular/core": "^14.2.1",
+    "@angular/forms": "^14.2.1",
+    "@angular/platform-browser": "^14.2.1",
+    "@angular/platform-browser-dynamic": "^14.2.1",
+    "@angular/router": "^14.2.1",
+    "@ng-bootstrap/ng-bootstrap": "^13.0.0",
+    "@popperjs/core": "^2.11.6",
+    "bootstrap": "^5.2.1",
+    "copyfiles": "^2.4.1",
+    "rxjs": "^7.5.6",
     "tslib": "^2.0.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "^0.11.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.0",
-    "@angular-eslint/builder": "13.1.0",
-    "@angular-eslint/eslint-plugin": "13.1.0",
-    "@angular-eslint/eslint-plugin-template": "13.1.0",
-    "@angular-eslint/schematics": "13.1.0",
-    "@angular-eslint/template-parser": "13.1.0",
-    "@angular/cli": "~13.3.0",
-    "@angular/compiler-cli": "~13.3.0",
-    "@angular/language-service": "~13.3.0",
+    "@angular-devkit/build-angular": "^14.2.2",
+    "@angular-eslint/builder": "14.0.4",
+    "@angular-eslint/eslint-plugin": "14.0.4",
+    "@angular-eslint/eslint-plugin-template": "14.0.4",
+    "@angular-eslint/schematics": "14.0.4",
+    "@angular-eslint/template-parser": "14.0.4",
+    "@angular/cli": "^14.2.2",
+    "@angular/compiler-cli": "^14.2.1",
+    "@angular/language-service": "^14.2.1",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "5.11.0",
-    "@typescript-eslint/parser": "5.11.0",
-    "eslint": "^8.2.0",
+    "@typescript-eslint/eslint-plugin": "^5.36.2",
+    "@typescript-eslint/parser": "^5.36.2",
+    "eslint": "^8.23.0",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
     "karma": "~6.3.16",
@@ -61,9 +62,8 @@
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^13.2.1",
-    "protractor": "~7.0.0",
+    "ng-packagr": "^14.2.1",
     "ts-node": "~7.0.0",
-    "typescript": "~4.5.5"
+    "typescript": "~4.8.3"
   }
 }
