@@ -14,9 +14,9 @@
     "tslib": "^2.0.0"
   },
   "peerDependencies": {
-    "@angular/common": ">=13.0.0",
-    "@angular/core": ">=13.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^11.0.0",
-    "bootstrap": "^4.6.0"
+    "@angular/common": ">=14.1.0",
+    "@angular/core": ">=14.1.0",
+    "@ng-bootstrap/ng-bootstrap": ">=13.0.0",
+    "bootstrap": ">=5.2.0"
   }
 }
